\name{SMM}
\alias{SMM}
\alias{sm}
\alias{smm}

\title{
	Spiral matrix generated by spiral mapping algorithm.
}
\description{
	We present an spiral mapping algorithm to generate spiral matrix in cancer research.
}
\usage{
	SSM(x)	
}
\arguments{
  \item{x}{
	A vector, each component of which has a tag name and is sorted in a certain way.
}
}
\author{
	Zhenzhang Li(design), Wen Luo(implementation, testing)
}
\examples{
	SMM(x)
}